requirements.txt
streamlit_chat==0.0.2.2
nltk==3.8.1
streamlit==1.19.0
langchain==0.0.129
python-dotenv~=1.0.0